Movement added for enemy
Basic movement added for enemy.

We may want to look at this a little but further on how we implement
this.

Right now it just has a max and min x and y pos and turns around once it
reaches them.

Note hard coded the max and min x and y in enemy creator.  Didn't want
to make too many changes if we were going to change how we did this.