Entity Holds Bounding Box and Enemies Die
The bounding boxes are held by every entity now, and are no longer
stored in components. This reduces duplication of data, plus it allows
us to tell if a bullet hits something that wouldn't necessarily have one
of the components with a bounding box

Enemies also get deleted when they are shot now.