Fixed info menu for linux
The problem was that windows doesn't care about capitalization in
filenames. So the original files were infomenuX.png but we loaded in
InfoMenuX.png which worked on windows but not linux. In the
compatibility update, I changed the file names to InfoMenuX.png. But,
since I made these changes on windows, git didn't think I changed the
file, so it was not changed on github.